Extract a code view into CodeView
