+++ CodeBreaker/CodeBreaker/CodeView.swift
@@ -1,0 +1,48 @@
+//
+//  CodeView.swift
+//  CodeBreaker
+//
+//  Created by Alexander Ostrovsky on 7/2/2026.
+//
+
+import SwiftUI
+
+struct CodeView: View {
+    // MARK: Data In
+    let code: Code
+    
+    // MARK: Data Shared with Me
+    @Binding var selection: Int
+    
+    // MARK: - Body
+    
+    var body: some View {
+        ForEach(code.pegs.indices, id: \.self) { index in
+            PegView(peg: code.pegs[index])
+                .padding(Selection.border)
+                .background {
+                    if selection == index, code.kind == .guess {
+                        Selection.shape
+                            .foregroundStyle(Selection.color)
+                    }
+                }
+                .onTapGesture {
+                    if code.kind == .guess {
+                        selection = index
+                    }
+                }
+        }
+    }
+    
+    struct Selection {
+        static let border: CGFloat = 5
+        static let cornerRadius: CGFloat = 10
+        static let color: Color = Color.gray(0.85)
+        static let shape = RoundedRectangle(cornerRadius: cornerRadius)
+    }
+}
+
+#Preview {
+    @Previewable @State var selection: Int = 0
+    CodeView(code: .init(kind: .guess, numberOfPegs: 4), selection: $selection)
+}
